eliminated matplotlib warning in kmeans
git-svn-id: https://svn.ais.uni-bonn.de/svn/repos/cuv/trunk@4904 9a9a52a9-7681-4bea-9213-143ca318e872

+++ examples/kmeans.py
@@ -41,6 +41,6 @@ if __name__ == "__main__":
     [clusters, indices]=kmeans(mnist,num_clusters,10)
 
     for i in xrange(num_clusters):
-        plt.subplot(np.ceil(num_clusters/5),5,i+1)
+        plt.subplot(int(np.ceil(num_clusters/5)),5,i+1)
         plt.imshow(clusters[:,i].reshape(28,28))
     plt.show()
